Moves the factories module to the search_client.test module
