--- search_client/factories/__init__.py
@@ -1,3 +1,0 @@
-from search_client.factories.learning_material import generate_nl_material, generate_material
-from search_client.factories.research_product import generate_nl_product, generate_product
-from search_client.factories.projects import generate_project
