+++ tests/test_backward_compatability.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 import json
 
 from tests.base import SearchClientTestCase
-from search_client.factories import generate_nl_material, generate_material, generate_nl_product, generate_product
+from search_client.test.factories import generate_nl_material, generate_material, generate_nl_product, generate_product
 from search_client.serializers.products import LearningMaterial, ResearchProduct
